Extracted default mind maps and options
